PBI-2360 PBI-2379 Criteria exception and handling (#500)
[PBI-2360]
[PBI-2379]

Did these together because they compliment each other.

Throwing an exception when we are not handling a targeting field is a good idea because otherwise it just goes through as false and silently fails. It makes us harder to evaluate what is happening if that happens.

Catching an exception in the loop where this happens will allow us to know what interaction is causing it. Also, it won't stop evaluating other options if this happens. We'll just log an error at that point.