PBI-5416 Launch event (#568)
* PBI-5416 Launch event
[PBI-5416]
Turns out onDestroy just doesn't seem to get called very often. Maybe when a garbage collect happens but not when forcefully closing the app.
And it also turns out that onStop doesn't get called when forcefully closing. But it does when backgrounding.
I also was wrong about rotation. It doesn't call our onStart method again, so it wouldn't call launch again.
However, I still think the correct course of action is for the SDK to call launch once per initialization.
Doing it this way will actually take the app's lead in determining if a app "launch" has occurred again, and not just a resume (from background).

* ktlint